% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1map_stegomyia_indices.R
\name{get_1map_stegomyia_indices}
\alias{get_1map_stegomyia_indices}
\title{get_1map_stegomyia_indices}
\usage{
get_1map_stegomyia_indices(df, m1 = shm1, m0 = shm2)
}
\arguments{
\item{df}{the dataframe with information processed by the function
"get_stegomyia_indices_by_type_of_study_and_geo_is" or
"get_stegomyia_indices_by_type_of_study_and_star_date".}

\item{m1}{is a shapefile"ejercicio_sectores_hermosillo.shp" with geographic
information of hermosillo city for join dataframe with information of
indices stegomyia}

\item{m0}{is a shapefile "ejercicio_sectores_hermosillo2.shp" with limit of
sectors of hermosillo city. This shapefile use to fution with m1 and give
limits of the sectors and the hermosillo city

  print(maps$HI)  # Print the map for the House Index, the Container Index 
  or the Breteau Index}
}
\value{
one map representing the different Stegomyia indices.
}
\description{
This function generates maps for three Stegomyia indices (Entomological Risk
Indices) of dataframe make to funtion
get_stegomyia_indices_by_type_of_study_and_geo_is.
This data frame containing information about sectors and their corresponding
indices.
}
\examples{

Sindx_map <- get_1map_stegomyia_indices(df_for_map)

Sindx_map

}
