% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stegomyia_indices_by_type_of_study_and_ew.R
\name{get_stegomyia_indices_by_type_of_study_and_ew}
\alias{get_stegomyia_indices_by_type_of_study_and_ew}
\title{Get stegomyia indices by type of study and star date}
\usage{
get_stegomyia_indices_by_type_of_study_and_ew(
  df,
  st = "Verificacion",
  ew = "1"
)
}
\arguments{
\item{df}{the dataframe with information processed by the function
"clean_raw_data".}

\item{st}{a string of type of study selected, "Encuesta" or "Verificacion". By
default, it is set to "Verificacion".}

\item{ew}{number of the epidemiological week used to calculate the stegomyia indices. By default,
it is set to "1".}
}
\value{
The dataframe with Stegomyia indices of the selection of type of study and
date.
}
\description{
The Stegomyia indices are calculated for each sampling. Once the data have
been loaded with function "load_rwa_data" and changed type data of function
"clean_raw_data", select study type and date variable using the following
 formulas:
      - Container Index(CI): (number of infected containers/total number of
      containers) * 100.
      - House Index (HI): (number of infected houses /total number of
      houses) * 100.
      - Breteau Index (BI): (number of positive containers/number of houses
      explored) * 100.
}
\examples{

df_stegomyia_sd <- get_stegomyia_indices_by_type_of_study_and_ew(df_indx,
                                                     st = "Verificacion",
  
                                                     ew = "2")
head(df_stegomyia_sd)

}
