% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stegomyia_indices_by_type_of_study_and_geo_is.R
\name{get_stegomyia_indices_by_type_of_study_and_geo_is}
\alias{get_stegomyia_indices_by_type_of_study_and_geo_is}
\title{Get stegomyia indices by type of study and geographic variable}
\usage{
get_stegomyia_indices_by_type_of_study_and_geo_is(
  df,
  st = "Verificacion",
  var,
  path_out = "data-raw/statusindicesector.csv"
)
}
\arguments{
\item{df}{The dataframe with entomology information processed by the function
"clean_raw_data".}

\item{st}{a string of type of study selected, "Encuesta" or "Verificacion". By
default, it is set to "Verificacion".}

\item{var}{receive the "Sector" number, which is the geographic variable 
used to calculate Stegomyia indexes.}

\item{path_out}{a string with path for the data of stegomyia indices and status. By
default, it is set to "data-raw/statusindicesector.csv".}
}
\value{
A dataframe with stegomyia indices and status of the selection of type of
study and geographic variable of sector.
}
\description{
The Stegomyia indices are calculated for each sampling. Once the data have
been loaded with the function "load_raw_data" and processed with the function
"clean_raw_data", you can select the study type and geographic variable using the
following formulas:
- Container Index (CI): (number of infected containers / total number of containers) * 100.
- House Index (HI): (number of infected houses / total number of houses) * 100.
- Breteau Index (BI): (number of positive containers / number of houses explored) * 100.
}
\examples{

path_index_table = c(system.file("extdata",
                                 "statusindicesector.csv",
                                  package = "rStegomyia"
                                  )
                     )


df_stegomyia <- get_stegomyia_indices_by_type_of_study_and_geo_is(df_indx,
                                                   st = "Verificacion",
                                                   var ="400",
                                                   path_out = path_index_table
  )

head(df_stegomyia)

}
