% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stegomyia_indices_by_type_of_study_and_location.R
\name{get_stegomyia_indices_by_type_of_study_and_loc}
\alias{get_stegomyia_indices_by_type_of_study_and_loc}
\title{Get stegomyia indices by type of study and geographic variable}
\usage{
get_stegomyia_indices_by_type_of_study_and_loc(
  df,
  st = "Verificacion",
  var,
  path_out = "data-raw/statusindicesloc.csv"
)
}
\arguments{
\item{df}{The dataframe with entomology information processed by the function
"clean_raw_data".}

\item{st}{a string of type of study selected, "Encuesta" or "Verificacion". By
default, it is set to "Verificacion".}

\item{var}{receive string of the "Location", which is the geographic variable 
used to calculate Stegomyia indexes.}

\item{path_out}{a string with path for the data of stegomyia indices and status. By
default, it is set to "data-raw/statusindicesloc.csv".}
}
\value{
A dataframe with stegomyia indices and status of the selection of type of
study and geographic variable of sector.
}
\description{
The Stegomyia indices are calculated for each sampling. Once the data have
been loaded with the function "load_raw_data" and processed with the function
"clean_raw_data", you can select the study type and geographic variable using the
following formulas:
- Container Index (CI): (number of infected containers / total number of containers) * 100.
- House Index (HI): (number of infected houses / total number of houses) * 100.
- Breteau Index (BI): (number of positive containers / number of houses explored) * 100.
}
\examples{

path_index_table = c(system.file("extdata",
                                 "statusindicesloc.csv",
                                  package = "rStegomyia"
                                  )
                     )


df_stegomyia_l <- get_stegomyia_indices_by_type_of_study_and_loc(df_indx,
                                                   st = "Verificacion",
                                                   var =c("HERMOSILLO",
                                                    "CIUDAD_OBREGÓN"),
                                                   path_out = path_index_table
               )

head(df_stegomyia_l)

}
