% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_stegomyia_indices_by_type_of_study_star_date_and_geo.R
\name{get_stegomyia_indices_by_type_of_study_star_date_and_geo}
\alias{get_stegomyia_indices_by_type_of_study_star_date_and_geo}
\title{Get stegomyia indices by type of study, date and geo}
\usage{
get_stegomyia_indices_by_type_of_study_star_date_and_geo(
  df,
  st = "Verificacion",
  date = "2021/01/07",
  var
)
}
\arguments{
\item{df}{The dataframe with entomology information processed by the function
"clean_raw_data".}

\item{st}{a string of type of study selected, "Encuesta" or "Verificacion". By
default, it is set to "Verificacion".}

\item{date}{a string with the date used to calculate the stegomyia indices. By default,
it is set to "2021/01/07".}

\item{var}{receive the "Sector" number, which is the geographic variable 
used to calculate Stegomyia indexes.}
}
\value{
The dataframe with stegomyia indices of the selection of type of study date,
and geographic variable of sector.
}
\description{
The Stegomyia indices are calculated for each sampling. Once the data have
been loaded with function "load_rwa_data" and changed type data of function
"clean_raw_data", select study type, date and geographic variable using the
following formulas:
      - Container Index (CI): (number of infected containers/total number of
      containers) * 100.
      - House Index (HI): (number of infected houses /total number of
      houses) * 100.
      - Breteau Index (BI): (number of positive containers/number of houses
      explored) * 100.
}
\examples{

df_dg <- get_stegomyia_indices_by_type_of_study_star_date_and_geo(df_indx,
st = "Encuesta",
date = "2021/01/06",
var = "390")

df_dg

}
