% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_typology_container_by_type_of_study_and_location.R
\name{get_typology_container_by_type_of_study_and_location}
\alias{get_typology_container_by_type_of_study_and_location}
\title{Positive Typology Container by type of study and location}
\usage{
get_typology_container_by_type_of_study_and_location(
  df,
  st = "Verificacion",
  var = c("CIUDAD_OBREGÓN", "HERMOSILLO")
)
}
\arguments{
\item{df}{The dataframe with entomology information processed by the function
"clean_raw_data".}

\item{st}{a string of type of study selected, "Encuesta" or "Verificacion". By
default, it is set to "Verificacion".}

\item{var}{receive string of the "Location", which is the geographic variable 
used to calculate Positive Typology Container.}
}
\value{
The dataframe with location of the selection of type of study and
geographic variable of location.
}
\description{
Complementary to the Stegomyia indices, the type of positive container is
 determined to establish the preference of mosquitoes in the use of
 breeding sites.
}
\examples{

df_tc <-get_typology_container_by_type_of_study_and_location(df_typ,
                                                             st = "Verificacion",
                                                             "CIUDAD_OBREGÓN"
)
head(df_tc)

}
