% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_data.R
\name{load_raw_data}
\alias{load_raw_data}
\title{Load raw data  of entomology study}
\usage{
load_raw_data(
  path,
  col_name = c("Tipo de Estudio", "Jurisdiccion", "Localidad", "Sector",
    "Fecha de Inicio", "Semana Epidemiologica", "Casas Revisadas", "Casas Positivas",
    "Total de Recipientes con Agua", "Total de Recipientes Positivos")
)
}
\arguments{
\item{path}{a string with raw data path of .txt file, of activity "Estudios
entomológicos en fase larval y pupal " of platafor "Vigilancia Entomológica
y Control Integral del Vector".}

\item{col_name}{selected variable names, as deemed necessary to analyze.}
}
\value{
A data frame with selected and format in variable of .txt file of plataform
"Vigilancia Entomológica y Control Integral del Vector" of activity "Estudio
Entomologico de fase larval y pupal".
}
\description{
Load_raw_data from a .txt file of the entomology study conducted on the
platform "Vigilancia Entomológica y Control Integral del Vector." Choose the
variables that will be used to calculate entomological risk indices. Fix
variable labels and eliminate spaces in the column names, replacing them
with underscores (_).If one of the columns labeled as Locality,
Municipality, or Jurisdiction is selected, two columns will be created from
the original one. the names of the new columns will be the key number of
the variable and the other with the name of the variable.
}
\examples{

path_raw_data <- system.file("extdata",
"estudio_entomologico1.txt",
package = "rStegomyia"
)

df_load_raw <- load_raw_data(path_raw_data)

head(df_load_raw)

}
